Implement JUnit testing for CommandResult
